packages/uniinfer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "uniinfer"
version = "0.4.8"
description = "Unified Inference API for LLM chat completions across 15+ providers with streaming, fallback strategies, and secure API key management"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT License"}
authors = [
  {name = "Han Woo", email = "dev@skale.dev"}
]
keywords = ["llm", "ai", "inference", "openai", "anthropic", "gemini", "mistral", "chat", "completion", "streaming", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests>=2.25.0",
    "python-dotenv>=1.0.0",
    "openai",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic",
    "python-multipart",
    "slowapi",
    "httpx",
    "zai-sdk>=0.2.2",
    "importlib-metadata; python_version<'3.8'",
    "credgoo @ git+https://github.com/devskale/python-openutils.git#subdirectory=packages/credgoo"
]

[project.optional-dependencies]
gemini = ["google-genai>=1.38.0"]
anthropic = ["anthropic>=0.25.0"]
mistral = ["mistralai>=0.4.0"]
cohere = ["cohere>=4.0.0"]
huggingface = ["huggingface-hub>=0.20.0"]
groq = ["groq>=0.11.0"]
ai21 = ["ai21>=2.1.0"]
all = [
    "google-genai>=1.38.0",
    "anthropic>=0.25.0",
    "mistralai>=0.4.0",
    "cohere>=4.0.0",
    "huggingface-hub>=0.20.0",
    "groq>=0.11.0",
    "ai21>=2.1.0"
]

[project.urls]
Homepage = "https://github.com/devskale/python-openutils"

[project.scripts]
uniinfer = "uniinfer.uniinfer_cli:main"
uniioai-proxy = "uniinfer.uniioai_proxy:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["uniinfer*"]

[tool.setuptools.package-data]
uniinfer = ["*.py", "examples/*.py", "examples/webdemo/*"]
